combine kernel: fx SharedAllocator + prune ep8 tuning
