+++ kernels/communication_ops_utils.py
@@ -6,7 +6,16 @@
 These wrap LLVM-dialect global memory ops with explicit memory ordering and
 syncscope -- which the high-level FlyDSL APIs (buffer_ops / Pointer) do not
 expose -- so dispatch/combine can publish and observe data across cards.
+
+Also hosts :class:`GeometryTuningTable`, the per-shape launch-geometry lookup
+shared by the dispatch/combine ops.
 """
+
+from __future__ import annotations
+
+import json
+from dataclasses import dataclass, field
+from typing import Dict, Tuple
 
 from flydsl._mlir import ir
 from flydsl._mlir.dialects import llvm as _llvm_d
@@ -18,6 +27,7 @@ __all__ = [
     "fence_system_acquire",
     "load_i64_global",
     "atomic_add_global_at",
+    "GeometryTuningTable",
 ]
 
 
@@ -63,3 +73,67 @@ def atomic_add_global_at(addr_i64, val):
     """Monotonic global ``atomic fetch-and-add``; returns the old value."""
     ptr = _to_ptr_global(addr_i64)
     return _llvm_d.AtomicRMWOp(_llvm_d.AtomicBinOp.add, ptr, arith.unwrap(val), _llvm_d.AtomicOrdering.monotonic).res
+
+
+@dataclass
+class GeometryTuningTable:
+    """Per-shape token-count -> (block_num, warp_num_per_block) lookup; rounds up
+    to the smallest bucket >= count (largest on overflow, mori parity)."""
+
+    dispatch: Dict[int, Tuple[int, int]] = field(default_factory=dict)
+    combine: Dict[int, Tuple[int, int]] = field(default_factory=dict)
+
+    def __post_init__(self):
+        for phase, tbl in (("dispatch", self.dispatch), ("combine", self.combine)):
+            for n_tok, (bn, wpb) in tbl.items():
+                if bn <= 0 or wpb <= 0:
+                    raise ValueError(
+                        f"GeometryTuningTable.{phase}[{n_tok}] must be positive, "
+                        f"got block_num={bn}, warp_num_per_block={wpb}"
+                    )
+
+    @classmethod
+    def from_tuning_file(
+        cls, path, *, dtype, hidden_dim, zero_copy, topk=None, local_expert_num=None, combine_dtype="bf16"
+    ):
+        """Build a per-op table from a multi-shape tuning JSON, filtered to this
+        op's shape; empty table => cfg defaults."""
+        with open(path, "r", encoding="utf-8") as f:
+            raw = json.load(f)
+
+        def _match(r, want_dtype, need_zc):
+            if r.get("dtype") != want_dtype or int(r.get("hidden_dim", -1)) != hidden_dim:
+                return False
+            if topk is not None and "topk" in r and int(r["topk"]) != topk:
+                return False
+            if (
+                local_expert_num is not None
+                and "local_expert_num" in r
+                and int(r["local_expert_num"]) != local_expert_num
+            ):
+                return False
+            if need_zc and bool(r.get("zero_copy", False)) != bool(zero_copy):
+                return False
+            return True
+
+        def _build(rules, want_dtype, need_zc):
+            return {
+                int(r["num_tokens"]): (int(r["block_num"]), int(r["warp_num_per_block"]))
+                for r in rules
+                if _match(r, want_dtype, need_zc)
+            }
+
+        return cls(
+            dispatch=_build(raw.get("dispatch", []), dtype, need_zc=False),
+            combine=_build(raw.get("combine", []), combine_dtype, need_zc=True),
+        )
+
+    def lookup(self, phase, num_tokens):
+        """Smallest bucket >= num_tokens (largest on overflow); None if empty."""
+        tbl = self.dispatch if phase == "dispatch" else self.combine
+        if not tbl:
+            return None
+        if num_tokens in tbl:
+            return tbl[num_tokens]
+        candidates = [k for k in tbl if k >= num_tokens]
+        return tbl[min(candidates)] if candidates else tbl[max(tbl)]
